:sparkles: Updated vendor URL and package version.
In this commit, the vendor URL for the gem 'active_merchant_pay2go' has been updated across multiple payment modes (production, development, and test) in the 'pay2go.rb' file to reflect a change in the vendor's official site. Similarly, the vendor's URL in the 'README.md' file was updated in relevant sections. Additionally, the gem's version number was incremented from "0.2.0" to "0.2.1" in 'version.rb'. These changes are necessary to keep up-to-date with the vendor's site updates and to reflect the successful incorporation of these changes into the library.